Reorganize app/lib
