+++ app/lib/selectors.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+module Selectors
+end
